requirements.txt
torch
torchvision
streamlit
pandas
scikit-learn
Pillow
